Secure the application

+++ app/api/issues/[id]/route.ts
@@ -1,50 +1,61 @@
+import authOptions from "@/app/auth/authOptions";
 import { issueSchema } from "@/app/validationSchemas";
 import prisma from "@/prisma/client";
 import delay from "delay";
+import { getServerSession } from "next-auth";
 import { NextRequest, NextResponse } from "next/server";
 
 
-export async function PATCH( 
-  request : NextRequest, 
-  { params }: { params: { id: string }} ) {
-    const body = await request.json();
-    const validation = issueSchema.safeParse(body);
+export async function PATCH(
+  request: NextRequest,
+  { params }: { params: { id: string } }) {
 
-    if(!validation.success)
-      return NextResponse.json(validation.error.format(), { status: 400 });
+  const session = await getServerSession(authOptions);
+  if (!session)
+    return NextResponse.json({}, { status: 401 });
 
-    const issue = await prisma.issue.findUnique({
-      where: { id: parseInt(params.id) }
-    });
+  const body = await request.json();
+  const validation = issueSchema.safeParse(body);
 
-    if(!issue)
-      return NextResponse.json({ error: "Invalid Issue" }, { status: 404 });
+  if (!validation.success)
+    return NextResponse.json(validation.error.format(), { status: 400 });
 
-    const updatedIssue = await prisma.issue.update({
-      where: { id: issue.id },
-      data: {
-        title: body.title,
-        description: body.description
-      }
-    });
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) }
+  });
 
-    return NextResponse.json(updatedIssue);
+  if (!issue)
+    return NextResponse.json({ error: "Invalid Issue" }, { status: 404 });
+
+  const updatedIssue = await prisma.issue.update({
+    where: { id: issue.id },
+    data: {
+      title: body.title,
+      description: body.description
+    }
+  });
+
+  return NextResponse.json(updatedIssue);
 }
 
-export async function DELETE( 
-  request : NextRequest, 
-  { params }: { params: { id: string }} ) {
+export async function DELETE(
+  request: NextRequest,
+  { params }: { params: { id: string } }) {
 
-    const issue = await prisma.issue.findUnique({
-      where: { id: parseInt(params.id) }
-    });
+  const session = await getServerSession(authOptions);
+  if (!session)
+    return NextResponse.json({}, { status: 401 });
 
-    if(!issue)
-      return NextResponse.json({ error: 'Invalid issue'}, { status: 404 });
+  const issue = await prisma.issue.findUnique({
+    where: { id: parseInt(params.id) }
+  });
 
-    await prisma.issue.delete({
-      where: { id: issue.id }
-    });
+  if (!issue)
+    return NextResponse.json({ error: 'Invalid issue' }, { status: 404 });
 
-    return NextResponse.json({});
-  }
+  await prisma.issue.delete({
+    where: { id: issue.id }
+  });
+
+  return NextResponse.json({});
+}
